+++ src/components/map/legend/mock.tsx
@@ -1,14 +1,4 @@
-import Icon from 'components/icon';
-import HEXAGON_SVG from 'svgs/map/hexagon.svg?sprite';
-
 const ITEMS = [
-  {
-    id: 'XXX',
-    name: 'Included areas',
-    icon: <Icon icon={HEXAGON_SVG} className="mt-1 h-3.5 w-3.5 text-purple-500" />,
-    description:
-      'Lorem ipsum dolor sit amet consectetur adipisicing elit. Natus minus eligendi doloremque unde, atque maxime dolore officiis quia architecto fugiat, dolorem animi vel! Velit minus facere maxime consequuntur iure. Nisi!',
-  },
   {
     id: 'YYY',
     name: 'All features',